refactor: move the Query base files to the Query directory
