+++ clients/java/src/main/java/io/camunda/zeebe/client/api/search/filter/DecisionDefinitionFilter.java
@@ -15,7 +15,7 @@
  */
 package io.camunda.zeebe.client.api.search.filter;
 
-import io.camunda.zeebe.client.api.search.TypedSearchQueryRequest.SearchRequestFilter;
+import io.camunda.zeebe.client.api.search.query.TypedSearchQueryRequest.SearchRequestFilter;
 
 public interface DecisionDefinitionFilter extends SearchRequestFilter {
