+++ clients/java/src/main/java/io/camunda/zeebe/client/api/search/query/DecisionRequirementsQuery.java
@@ -16,8 +16,6 @@
 
 package io.camunda.zeebe.client.api.search.query;
 
-import io.camunda.zeebe.client.api.search.FinalSearchQueryStep;
-import io.camunda.zeebe.client.api.search.TypedSearchQueryRequest;
 import io.camunda.zeebe.client.api.search.filter.DecisionRequirementsFilter;
 import io.camunda.zeebe.client.api.search.response.DecisionRequirements;
 import io.camunda.zeebe.client.api.search.sort.DecisionRequirementsSort;
