+++ clients/java/src/main/java/io/camunda/zeebe/client/api/search/query/SearchRequestBuilders.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.zeebe.client.api.search;
+package io.camunda.zeebe.client.api.search.query;
 
 import io.camunda.zeebe.client.api.search.filter.DecisionDefinitionFilter;
 import io.camunda.zeebe.client.api.search.filter.DecisionRequirementsFilter;
@@ -30,7 +30,7 @@ import io.camunda.zeebe.client.impl.search.filter.DecisionRequirementsFilterImpl
 import io.camunda.zeebe.client.impl.search.sort.DecisionRequirementsSortImpl;
 import io.camunda.zeebe.client.impl.search.filter.ProcessInstanceFilterImpl;
 import io.camunda.zeebe.client.impl.search.sort.ProcessInstanceSortImpl;
-import io.camunda.zeebe.client.impl.search.SearchRequestPageImpl;
+import io.camunda.zeebe.client.impl.search.query.SearchRequestPageImpl;
 import io.camunda.zeebe.client.impl.search.filter.UserTaskFilterImpl;
 import io.camunda.zeebe.client.impl.search.sort.UserTaskSortImpl;
 import io.camunda.zeebe.client.impl.search.filter.VariableValueFilterImpl;
