+++ clients/java/src/main/java/io/camunda/zeebe/client/api/search/query/SearchRequestPage.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.zeebe.client.api.search;
+package io.camunda.zeebe.client.api.search.query;
 
 import java.util.List;
