+++ clients/java/src/main/java/io/camunda/zeebe/client/api/search/sort/DecisionDefinitionSort.java
@@ -15,7 +15,7 @@
  */
 package io.camunda.zeebe.client.api.search.sort;
 
-import io.camunda.zeebe.client.api.search.TypedSearchQueryRequest.SearchRequestSort;
+import io.camunda.zeebe.client.api.search.query.TypedSearchQueryRequest.SearchRequestSort;
 
 public interface DecisionDefinitionSort extends SearchRequestSort<DecisionDefinitionSort> {
