+++ clients/java/src/main/java/io/camunda/zeebe/client/api/search/sort/UserTaskSort.java
@@ -16,7 +16,7 @@
 
 package io.camunda.zeebe.client.api.search.sort;
 
-import io.camunda.zeebe.client.api.search.TypedSearchQueryRequest.SearchRequestSort;
+import io.camunda.zeebe.client.api.search.query.TypedSearchQueryRequest.SearchRequestSort;
 
 public interface UserTaskSort extends SearchRequestSort<UserTaskSort> {
