+++ clients/java/src/main/java/io/camunda/zeebe/client/impl/search/filter/DecisionRequirementsFilterImpl.java
@@ -16,7 +16,7 @@
 package io.camunda.zeebe.client.impl.search.filter;
 
 import io.camunda.zeebe.client.api.search.filter.DecisionRequirementsFilter;
-import io.camunda.zeebe.client.impl.search.TypedSearchRequestPropertyProvider;
+import io.camunda.zeebe.client.impl.search.query.TypedSearchRequestPropertyProvider;
 import io.camunda.zeebe.client.protocol.rest.DecisionRequirementsFilterRequest;
 
 public class DecisionRequirementsFilterImpl
