+++ clients/java/src/main/java/io/camunda/zeebe/client/impl/search/filter/VariableValueFilterImpl.java
@@ -16,7 +16,7 @@
 package io.camunda.zeebe.client.impl.search.filter;
 
 import io.camunda.zeebe.client.api.search.filter.VariableValueFilter;
-import io.camunda.zeebe.client.impl.search.TypedSearchRequestPropertyProvider;
+import io.camunda.zeebe.client.impl.search.query.TypedSearchRequestPropertyProvider;
 import io.camunda.zeebe.client.protocol.rest.VariableValueFilterRequest;
 
 public class VariableValueFilterImpl
