+++ clients/java/src/main/java/io/camunda/zeebe/client/impl/search/query/DecisionRequirementsQueryImpl.java
@@ -16,23 +16,20 @@
 
 package io.camunda.zeebe.client.impl.search.query;
 
-import static io.camunda.zeebe.client.api.search.SearchRequestBuilders.decisionRequirementsSort;
-import static io.camunda.zeebe.client.api.search.SearchRequestBuilders.searchRequestPage;
+import static io.camunda.zeebe.client.api.search.query.SearchRequestBuilders.decisionRequirementsSort;
+import static io.camunda.zeebe.client.api.search.query.SearchRequestBuilders.searchRequestPage;
 
 import io.camunda.zeebe.client.api.JsonMapper;
 import io.camunda.zeebe.client.api.search.filter.DecisionRequirementsFilter;
 import io.camunda.zeebe.client.api.search.query.DecisionRequirementsQuery;
 import io.camunda.zeebe.client.api.search.sort.DecisionRequirementsSort;
-import io.camunda.zeebe.client.api.search.FinalSearchQueryStep;
-import io.camunda.zeebe.client.api.search.SearchRequestBuilders;
-import io.camunda.zeebe.client.api.search.SearchRequestPage;
+import io.camunda.zeebe.client.api.search.query.FinalSearchQueryStep;
+import io.camunda.zeebe.client.api.search.query.SearchRequestBuilders;
+import io.camunda.zeebe.client.api.search.query.SearchRequestPage;
 import io.camunda.zeebe.client.api.search.response.DecisionRequirements;
 import io.camunda.zeebe.client.api.search.response.SearchQueryResponse;
 import io.camunda.zeebe.client.impl.http.HttpClient;
 import io.camunda.zeebe.client.impl.http.HttpZeebeFuture;
-import io.camunda.zeebe.client.impl.search.SearchRequestPageImpl;
-import io.camunda.zeebe.client.impl.search.SearchResponseMapper;
-import io.camunda.zeebe.client.impl.search.TypedSearchRequestPropertyProvider;
 import io.camunda.zeebe.client.impl.search.sort.DecisionRequirementsSortImpl;
 import io.camunda.zeebe.client.protocol.rest.DecisionRequirementsFilterRequest;
 import io.camunda.zeebe.client.protocol.rest.DecisionRequirementsSearchQueryRequest;
