+++ clients/java/src/main/java/io/camunda/zeebe/client/impl/search/query/SearchQuerySortBase.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.zeebe.client.impl.search;
+package io.camunda.zeebe.client.impl.search.query;
 
 import io.camunda.zeebe.client.protocol.rest.SearchQuerySortRequest;
 import java.util.ArrayList;
